feat: utils functions #20
