+++ next/pages/bio-networks/[network]/datasets.tsx
@@ -1,33 +1,14 @@
-import {
-  GetStaticPaths,
-  GetStaticProps,
-  GetStaticPropsContext,
-  InferGetStaticPropsType,
-} from "next";
+import { GetStaticPaths, GetStaticProps, InferGetStaticPropsType } from "next";
 import React from "react";
-import { NetworkParam } from ".";
-import { NETWORKS } from "../../../constants/networks";
+import { NetworkParam } from "../../../@types/network";
 import { NetworkProvider } from "../../../contexts/networkContext";
+import * as networkPages from "../../../utils/networkPages";
 import { NetworkDetailDatasets } from "../../../views/NetworkDetailView/NetworkDetailDatasets/networkDetailDatasets";
 
-export const getStaticPaths: GetStaticPaths = async () => {
-  return {
-    paths: NETWORKS.map((network) => ({
-      params: { network: network.path },
-    })),
-    fallback: false,
-  };
-};
+export const getStaticPaths: GetStaticPaths = networkPages.getStaticPaths;
 
-export const getStaticProps: GetStaticProps<NetworkParam> = async (
-  context: GetStaticPropsContext
-) => {
-  const { network: networkParam } = context.params ?? {};
-
-  const network = NETWORKS.find(({ path }) => path === networkParam)!;
-
-  return { props: { network } };
-};
+export const getStaticProps: GetStaticProps<NetworkParam> =
+  networkPages.getStaticProps;
 
 const Page = ({ network }: InferGetStaticPropsType<typeof getStaticProps>) => {
   return (
